requirements.txt
fastapi
motor
uvicorn
torch
numpy
scipy
sklearn
pillow
beanie
pydantic[dotenv]
python-multipart